FC5 compilation fixes that were sent to the bug tracker.
I did not include the codasrv fix that links it with voldump.client.o
because I'm not sure if the problem can be fixed by shuffling the
order in which we link libraries. Although there might be a circular
dependency since libvicedep is both a consumer and a provider of
functions. Maybe vicedep needs to be split up into vicedep-client and
vicedep-server.