Kick the FSO daemon when we hit yellow or red zones.
We can get stuck in a 'yellow' or 'red' zone state even when the CML
is empty. The fso_daemon thread only reclaims fsos once every 30
seconds, so it is possible that we fill up the cache quicker than the
fso daemon clear out cache space (because the of redzoning almost all
reclaims should be triggered by the fso daemon).