Allow partial commit for log-based resolution.
Log-based resolution was unable to correctly handle the following
sequence of operations,
    create a/x ; mv b/x b/y ; mv a/x b/x ; rm b/y

The result of this sequence is directory 'a' with 'create;mv x->b/x'
and directory 'b' with 'mv x->y;mv a/x->x;rm y'. The problem at this
point is that we cannot successfully resolve the rename in a, because
'b/x' already exists, and we can not resolve the second rename in b
because 'a/x' hasn't been created yet.

This patch allows resolution to commit all operations that could be
applied and fail the rest. Because we fall back on hinting the other
directory we can now successfully resolve the previous case. If we
start in directory a, we resolve 'create x', then move on to the
other directory which can complete and then we go back to a which
ends up only requiring a version vector sync. When we start in b, we
apply 'mv x->y', hint a which completes and then return to b and apply
the remaining 'rm y'.