Add SF_UNCACHED bit to va_flags in the getattr response.
This can be used by the kernel as a hint whether the file contents
still has to be fetched from the servers. On windows we can use this
to set the FILE_OFFLINE property in which case explorer will not try
to open the file to read icons and generate thumbnail images and such.