Fix 'fatal error -- fsobj::AttachChild: not dir' during recovery.
We would crash during recover when we found a directory that is both
marked as having a conflict and being a runt. Move the IsFake() test out
of HAVESTATUS (which really wants to test if the object happens to be a
runt or not) and then add IsFake exceptions in the few places where we
do need to make progress even for runt objects as objects in conflict
will never matriculate and we need to be able to reference them when
expanding or collapsing conflicts.