Try to reintegrate over conflicts with cfs forcereintegrate
Ignore the to_be_repaired flag whenever we are running a synchronous
reintegration. Unlike cfs preservelocal we do not try to mess with fids,
version vectors and such. This means that if there is a valid reason for
the conflict this new attempt will simply fail again, but we do not risk
corrupting or crashing the client.