Add support for newer POSIX.1 ustar format as well as 2 cpio variants.
We would write out a corrupt tar archive with CML data if any path
names exceeded 100 characters, this limit is inherent in the old
BSD4.3 tar format the client was using. This patch adds support for
a couple of other archive formats, each with their own limitations.

ustar format, which allows for pathnames up to 255 characters as long
as the final path components remains < 100 chars.

Added the older but POSIX.1 standard cpio 'odc' format, which can in
theory support up to 262143 character pathnames, in practice we hit
other PATH_MAX based limits before that. However this format has very
small limited inode and userid spaces. The uid space is not as much of
an issue, but the inode numbers are used to find hardlinks. So we also
introduce the SVR4 'newc' format, which uses 32-bit fields everywhere,
giving plenty of inode and uid space, but limits file sizes to 4GB.

Moved the code that handles the lower level writing of archive files
out of vol_cml.cc. It should get cleaned up more before we can use the
same code in other places that need to create archives (codadump2tar).