Avoid assertion when forced directory operations fail.
Asserting here seems to leave a directory in RVM without any content.
'seems' because it is unclear to me how not committing bad in-memory
state could corrupt RVM, which indicates that it must already be bad
before we even entered this function. If we hobble on, we at least
give PutObjects a chance to clean up the bad state and return an
error to the client instead of stopping the server.