Propagate const'ness of strings we are looking up.
Found some ugly things related to case insensitive lookups, use
strcasecmp instead of subtracting 'A' - 'a' for uppercase characters,
which only works for the ascii "a-z" characters and fails for umlauts
etc. At least strcasecmp can get a little closer to reality, although it
is broken as well once you consider that we may have a mix of UTF-8 and
other encodings. And definitely do not copy the found string back over
the one we are looking for since that may have been a constant string
value.